Used synchronized callbacks for PeerConnection
