+++ include/rtc/channel.hpp
@@ -54,25 +54,6 @@ protected:
 	virtual void triggerSent();
 
 private:
-	template <typename... P> class synchronized_callback {
-	public:
-		synchronized_callback &operator=(std::function<void(P...)> func) {
-			std::lock_guard<std::recursive_mutex> lock(mutex);
-			callback = func;
-			return *this;
-		}
-
-		void operator()(P... args) {
-			std::lock_guard<std::recursive_mutex> lock(mutex);
-			if (callback)
-				callback(args...);
-		}
-
-	private:
-		std::function<void(P...)> callback;
-		std::recursive_mutex mutex;
-	};
-
 	synchronized_callback<> mOpenCallback;
 	synchronized_callback<> mClosedCallback;
 	synchronized_callback<const string &> mErrorCallback;
